requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==4.0
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pycodestyle==2.13.0
python-dotenv==1.1.0
requests==2.32.3
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
